+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/SearchOptions.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import com.mapbox.geojson.BoundingBox
 import com.mapbox.geojson.Point
 import com.mapbox.search.Reserved.Flags.SBS
+import com.mapbox.search.Reserved.Flags.SEARCH_BOX
 import com.mapbox.search.RouteOptions.Deviation.SarType
 import com.mapbox.search.base.assertDebug
 import com.mapbox.search.base.core.CoreSearchOptions
@@ -74,25 +75,25 @@ public class SearchOptions @JvmOverloads public constructor(
     /**
      * Point for ETA calculation from it to search result.
      *
-     * Note: Supported for Single Box Search API only. Reserved for internal and special use.
+     * Note: Supported for Single Box Search and Search Box APIs only. Reserved for internal and special use.
      */
-    @Reserved(SBS)
+    @Reserved(SBS, SEARCH_BOX)
     public val origin: Point? = null,
 
     /**
      * Navigation options used for proper calculation of ETA and results ranking.
      *
-     * Note: Supported for Single Box Search API only. Reserved for internal and special use.
+     * Note: Supported for Single Box Search and Search Box APIs only. Reserved for internal and special use.
      */
-    @Reserved(SBS)
+    @Reserved(SBS, SEARCH_BOX)
     public val navigationOptions: SearchNavigationOptions? = null,
 
     /**
      * Options to configure Route for search along the route functionality.
      *
-     * Note: Supported for Single Box Search API only. Reserved for internal and special use.
+     * Note: Supported for Single Box Search and Search Box APIs only. Reserved for internal and special use.
      */
-    @Reserved(SBS)
+    @Reserved(SBS, SEARCH_BOX)
     public val routeOptions: RouteOptions? = null,
 
     /**
@@ -116,14 +117,6 @@ public class SearchOptions @JvmOverloads public constructor(
      * Threshold specified in meters.
      */
     public val indexableRecordsDistanceThresholdMeters: Double? = null,
-
-    /**
-     * Besides the basic metadata attributes, developers can request additional
-     * attributes by setting attribute_sets parameter with attribute set values,
-     * for example &attribute_sets=basic,photos,visit.
-     * The requested metadata will be provided in metadata object in the response.
-     */
-    public val attributeSets: List<AttributeSet>? = null,
 ) : Parcelable {
 
     init {
@@ -152,7 +145,6 @@ public class SearchOptions @JvmOverloads public constructor(
         unsafeParameters: Map<String, String>? = this.unsafeParameters,
         ignoreIndexableRecords: Boolean = this.ignoreIndexableRecords,
         indexableRecordsDistanceThresholdMeters: Double? = this.indexableRecordsDistanceThresholdMeters,
-        attributeSets: List<AttributeSet>? = this.attributeSets,
     ): SearchOptions {
         return SearchOptions(
             proximity = proximity,
@@ -169,7 +161,6 @@ public class SearchOptions @JvmOverloads public constructor(
             unsafeParameters = unsafeParameters,
             ignoreIndexableRecords = ignoreIndexableRecords,
             indexableRecordsDistanceThresholdMeters = indexableRecordsDistanceThresholdMeters,
-            attributeSets = attributeSets,
         )
     }
 
@@ -203,7 +194,6 @@ public class SearchOptions @JvmOverloads public constructor(
         if (unsafeParameters != other.unsafeParameters) return false
         if (ignoreIndexableRecords != other.ignoreIndexableRecords) return false
         if (!indexableRecordsDistanceThresholdMeters.safeCompareTo(other.indexableRecordsDistanceThresholdMeters)) return false
-        if (attributeSets != other.attributeSets) return false
 
         return true
     }
@@ -226,7 +216,6 @@ public class SearchOptions @JvmOverloads public constructor(
         result = 31 * result + (unsafeParameters?.hashCode() ?: 0)
         result = 31 * result + ignoreIndexableRecords.hashCode()
         result = 31 * result + indexableRecordsDistanceThresholdMeters.hashCode()
-        result = 31 * result + (attributeSets?.hashCode() ?: 0)
         return result
     }
 
@@ -249,7 +238,6 @@ public class SearchOptions @JvmOverloads public constructor(
                 "unsafeParameters=$unsafeParameters, " +
                 "ignoreIndexableRecords=$ignoreIndexableRecords, " +
                 "indexableRecordsDistanceThresholdMeters=$indexableRecordsDistanceThresholdMeters" +
-                "attributeSets=$attributeSets" +
                 ")"
     }
 
@@ -290,7 +278,6 @@ public class SearchOptions @JvmOverloads public constructor(
             unsafeParameters = options.unsafeParameters
             ignoreIndexableRecords = options.ignoreIndexableRecords
             indexableRecordsDistanceThresholdMeters = options.indexableRecordsDistanceThresholdMeters
-            attributeSet = options.attributeSets
         }
 
         /**
@@ -359,17 +346,17 @@ public class SearchOptions @JvmOverloads public constructor(
         /**
          * Point for ETA calculation from it to search result.
          *
-         * Note: Supported for Single Box Search API only. Reserved for internal and special use.
+         * Note: Supported for Single Box Search and Search Box APIs only. Reserved for internal and special use.
          */
-        @Reserved(SBS)
+        @Reserved(SBS, SEARCH_BOX)
         public fun origin(origin: Point): Builder = apply { this.origin = origin }
 
         /**
          * Navigation options used for proper calculation of ETA and results ranking.
          *
-         * Note: Supported for Single Box Search API only. Reserved for internal and special use.
+         * Note: Supported for Single Box Search and Search Box APIs only. Reserved for internal and special use.
          */
-        @Reserved(SBS)
+        @Reserved(SBS, SEARCH_BOX)
         public fun navigationOptions(navigationOptions: SearchNavigationOptions): Builder = apply {
             this.navigationOptions = navigationOptions
         }
@@ -377,9 +364,9 @@ public class SearchOptions @JvmOverloads public constructor(
         /**
          * Options to configure Route for search along the route functionality.
          *
-         * Note: Supported for Single Box Search API only. Reserved for internal and special use.
+         * Note: Supported for Single Box Search and Search Box APIs only. Reserved for internal and special use.
          */
-        @Reserved(SBS)
+        @Reserved(SBS, SEARCH_BOX)
         public fun routeOptions(routeOptions: RouteOptions): Builder = apply { this.routeOptions = routeOptions }
 
         /**
@@ -410,16 +397,6 @@ public class SearchOptions @JvmOverloads public constructor(
         }
 
         /**
-         * Besides the basic metadata attributes, developers can request additional
-         * attributes by setting attribute_sets parameter with attribute set values,
-         * for example &attribute_sets=basic,photos,visit.
-         * The requested metadata will be provided in metadata object in the response.
-         */
-        public fun attributeSet(attributeSet: List<AttributeSet>?): Builder = apply {
-            this.attributeSet = attributeSet
-        }
-
-        /**
          * Create [SearchOptions] instance from builder data.
          */
         public fun build(): SearchOptions = SearchOptions(
@@ -437,7 +414,6 @@ public class SearchOptions @JvmOverloads public constructor(
             unsafeParameters = unsafeParameters,
             ignoreIndexableRecords = ignoreIndexableRecords,
             indexableRecordsDistanceThresholdMeters = indexableRecordsDistanceThresholdMeters,
-            attributeSets = attributeSet
         )
     }
 }
@@ -474,8 +450,7 @@ internal fun SearchOptions.mapToCore(): CoreSearchOptions = CoreSearchOptions(
     routeOptions?.route,
     routeOptions?.deviation?.sarType?.rawName,
     routeOptions?.timeDeviationMinutes,
-    unsafeParameters?.let { (it as? HashMap) ?: HashMap(it) },
-    attributeSets?.map { it.mapToCore() }
+    unsafeParameters?.let { (it as? HashMap) ?: HashMap(it) }
 )
 
 @JvmSynthetic
@@ -506,6 +481,5 @@ internal fun CoreSearchOptions.mapToPlatform(): SearchOptions = SearchOptions(
     } else null,
     unsafeParameters = addonAPI,
     ignoreIndexableRecords = ignoreUR,
-    indexableRecordsDistanceThresholdMeters = urDistanceThreshold,
-    attributeSets = attributeSets?.map { it.mapToPlatform() },
+    indexableRecordsDistanceThresholdMeters = urDistanceThreshold
 )